Finished project
- Finished add and remove from database functions and forms, but i am going to work on the css and style it better than its current state.

- Getting the add_object to db fucntion to work was easy but getting the delete function to work corrently took about 45 minutes, the documentation was not very clear in my opinion on this and i ended up getting the solution finally working from this source (https://stackoverflow.com/questions/51096776/how-to-delete-row-from-table-in-codeigniter) idk why its delete solution was so much different from all of the other methods i tried, i kept running into an error where there was no table selected (documentation on delete() doesnt state that it needs a table paremeter).

I completed the project in about 8 hours (actual coding time was 4.5-5 hours) had some personal things to do today which interrupted my work a few times but i got it done.

I am going to continue to work on the project and add delete links to each object rather than delete by id which is what the instructions asked for originally and clean up the css of the forms and make it look nicer.

Thank you for the code exercise it was fun learning a new technology, funnily enough i can report that the php frameworks are very similar to the django framework which is the one i have the most experience on.